# src/swagger-config.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CASMI APIs",
        "description": "CASMI: Cloud Agnostic Storage Management Interface",
        "version": "1.0"
    },
    "servers": [{ "url": "http://localhost:8000/api/v1" }],
    "paths": {
...
        "/spaces": {
            "get": {
                "summary": "Summary",
                "description": "Description",
                "tags": ["Spaces"],
                "security": [{ "Bearer token": [] }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "spaces": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": { "type": "integer" },
                                                    "name": { "type": "string" },
                                                    "providers": {
                                                        "type": "array",
                                                        "items": { "type": "strin" }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "message": { "type": "string" }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Summary",
                "description": "Description",
                "tags": ["Spaces"],
                "security": [{ "Bearer token": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": { "type": "string" },
                                    "provider_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "minItems": 1
                                    },
                                    "default_secrets": { "type": "boolean" },
                                    "secrets": { "type": "object" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created new space.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "space_id": { "type": "integer" }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Client error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "message": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "message": { "type": "string" }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/spaces/{id}": {
            "get": {
                "summary": "Summary",
                "description": "Description",
                "tags": ["Spaces"],
                "security": [{ "Bearer token": [] }],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "The space ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "space": {
                                            "type": "object",
                                            "properties": {
                                                "id": { "type": "integer" },
                                                "name": { "type": "string" },
                                                "providers": {
                                                    "type": "array",
                                                    "items": { "type": "strin" }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Client error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "message": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "message": { "type": "string" }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer token": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}